requirements.txt
asgiref==3.6.0
black==22.6.0
boto3==1.26.84
botocore==1.29.84
certifi==2022.12.7
click==8.0.4
colorama==0.4.6
Django==4.1.7
django-prometheus==2.2.0
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
mypy-extensions==0.4.3
pathspec==0.10.3
Pillow==9.4.0
platformdirs==2.5.2
prometheus-client==0.16.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
wincertstore==0.2
